Conflicts:
	3d.cc
	3d.h
	distance_est.cc
	getcolor.cc
	getparams.c
	mandelbulb_dist_est.cc
	raymarching.cc
	renderer.cc

PLEASE CHECK THAT THIS COMMIT WORKS ON OUR GPU'S

+++ camera.h
@@ -21,15 +21,15 @@
 #ifndef _CAMERA_H
 #define _CAMERA_H
 
-typedef struct 
+typedef struct
 {
-	double camPos[3];
-	double camTarget[3];
-	double camUp[3];
-	double fov;
-	double matModelView[16];
-	double matProjection[16];
-	double matInvProjModel[16]; 
+	float camPos[3];
+	float camTarget[3];
+	float camUp[3];
+	float fov;
+	float matModelView[16];
+	float matProjection[16];
+	float matInvProjModel[16]; 
 	int    viewport[4];
 } CameraParams;
